+++ docs/changelog.md
@@ -34,6 +34,10 @@
 
 ## 更新说明
 
+1.7.54 fix: fix stock_a_lg_indicator interface
+
+    1. 修复 stock_a_lg_indicator 接口
+
 1.7.53 add: add stock_buffett_index_lg interface
 
     1. 新增 stock_buffett_index_lg 接口
@@ -1323,6 +1327,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.7.54 fix: fix stock_a_lg_indicator interface
 
 1.7.53 add: add stock_buffett_index_lg interface